+++ api/Repositories/BidRepo/IBidRepository.cs
@@ -5,9 +5,9 @@ namespace api.Repositories.BidRepo
     public interface IBidRepository
     {
         Task<IEnumerable<Bid>> GetAllAsync();
-        Task<Bid?> GetByIdAsync(int id);
+        Task<Bid?> GetByIdAsync(long id);
         Task<Bid> AddBidAsync(Bid bid);
         Task<Bid> UpdateBidAsync(Bid bid);
-        Task DeleteBidAsync(int id);
+        Task DeleteBidAsync(long id);
     }
 }
