+++ api/Repositories/UserRepo/IUserRepository.cs
@@ -6,12 +6,12 @@ namespace api.Repositories.UserRepo
     {
         Task<IEnumerable<User>> GetAllAsync();
 
-        Task<User?> GetByIdAsync(int id);
+        Task<User?> GetByIdAsync(long id);
 
         Task<User> AddUserAsync(User user);
 
         Task<User> UpdateUserAsync(User user);
 
-        Task DeleteUserAsync(int id);
+        Task DeleteUserAsync(long id);
     }
 }
